api/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
mistralai==0.0.7
sounddevice==0.4.6
wave==0.0.2
numpy==1.26.2
requests==2.31.0
PyPDF2==3.0.1
langdetect==1.0.9